Split commons eclipse pde to remove UI dependency (#70)
* update due to split of org.eclipse.gemoc.commons.eclipse.pde

* reduce dependencies

use GemocPlugin in commons.eclipse instead of commons.eclipse.pde
