+++ ccsljava_engine/plugins/org.eclipse.gemoc.execution.concurrent.ccsljavaengine.extensions.timesquare/src/main/java/org/eclipse/gemoc/execution/concurrent/ccsljavaengine/extensions/timesquare/Activator.java
@@ -14,7 +14,7 @@ package org.eclipse.gemoc.execution.concurrent.ccsljavaengine.extensions.timesqu
 import org.eclipse.gemoc.commons.eclipse.logging.backends.DefaultLoggingBackend;
 import org.eclipse.gemoc.commons.eclipse.messagingsystem.api.MessagingSystem;
 import org.eclipse.gemoc.commons.eclipse.messagingsystem.api.MessagingSystemManager;
-import org.eclipse.gemoc.commons.eclipse.pde.GemocPlugin;
+import org.eclipse.gemoc.commons.eclipse.GemocPlugin;
 import org.osgi.framework.BundleContext;
