+++ ccsljava_engine/plugins/org.eclipse.gemoc.execution.concurrent.ccsljavaengine/META-INF/MANIFEST.MF
@@ -8,7 +8,6 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.gemoc.execution.concurrent.ccsljavaxdsml.api;bundle-version="1.0.0",
  org.eclipse.gemoc.executionframework.engine;bundle-version="3.0.0",
  org.eclipse.gemoc.xdsmlframework.api;bundle-version="3.0.0",
- org.eclipse.gemoc.commons.eclipse.pde;bundle-version="0.1.0",
  org.eclipse.emf.transaction;bundle-version="1.8.0",
  com.google.guava;bundle-version="15.0.0",
  org.eclipse.emf.compare,
@@ -17,7 +16,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.gemoc.execution.concurrent.ccsljavaengine.mse.model;bundle-version="0.1.0",
  org.eclipse.gemoc.trace.gemoc.api,
  org.eclipse.debug.core,
- org.eclipse.gemoc.commons.eclipse,
+ org.eclipse.gemoc.commons.eclipse;bundle-version="3.6.0",
  org.eclipse.gemoc.trace.commons;bundle-version="1.0.0",
  fr.inria.diverse.k3.al.annotationprocessor.plugin,
  fr.inria.aoste.timesquare.instantrelation,
@@ -28,6 +27,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.gemoc.execution.concurrent.ccsljavaengine.extensions.k3;bundle-version="2.3.0",
  org.apache.commons.lang3;bundle-version="3.1.0",
  org.eclipse.gemoc.execution.moccml.testscenariolang.model;bundle-version="1.0.0",
+ org.eclipse.pde.core,
  org.eclipse.jdt.core,
  org.eclipse.gemoc.execution.moccml.testscenariolang.xtext;bundle-version="1.0.0",
  org.eclipse.gemoc.execution.moccml.testscenariolang.xtext.ui;bundle-version="1.0.0",
