+++ ccsljava_xdsml/plugins/org.eclipse.gemoc.execution.concurrent.ccsljavaxdsml.ui/META-INF/MANIFEST.MF
@@ -25,6 +25,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.gemoc.executionframework.ui,
  fr.inria.aoste.timesquare.ccslkernel.model;bundle-version="1.0.0",
  org.eclipse.gemoc.commons.eclipse.pde,
+ org.eclipse.gemoc.commons.eclipse.pde.ui,
  org.eclipse.gemoc.commons.eclipse.jdt,
  org.eclipse.gemoc.commons.eclipse.ui,
  org.eclipse.gemoc.execution.concurrent.ccsljavaengine.extensions.k3;bundle-version="0.1.0",
